requirements.txt
flask
gunicorn
gpytranslate
pytgcalls==3.0.0.dev22
openai
uvloop
ffmpeg-python
yt-dlp
youtube-search
youtube-search-python
APScheduler==3.7.0
aiohttp
aiofiles
bs4
emoji==1.6.0
gitpython
httpx[http2]
humanize
gTTs
lottie
httpx
heroku3
requests
motor==3.0.0
opencv-python-headless
psutil
pyrogram
pymongo
pyromod==1.1.0
pytube
python-dotenv
pymediainfo
SpeechRecognition
prettytable
git+https://github.com/Itz-fork/py-trans.git
reportlab
svglib
speedtest-cli==2.1.3
tgcrypto 
textblob
tracemoepy
urllib3
uvloop
telegraph
uwuify
wget
pyaes==1.6.1
pysocks==1.7.1
aiosqlite>=0.16.0,<0.18.0
Pillow
git+https://github.com/Itz-fork/py-extract.git
ubotlibs